Restore ColumnSum SIMD specializations - extract Sum3x3/Sum5x5 to box_filter_2d.simd.hpp - revert APPLY_SCALING_WITH_SHIFT in columnSum. - ColumnSum 16u and 16s template removed. - filter engine inner loop condition checks avoided. - gcc loop align - minimize 32f box filter regression on Intel IceLake
